Refact: volume data structures moved to relevant module.
