+++ src/adf_str.h
@@ -34,32 +34,6 @@
 #include"adf_blk.h"
 #include"adf_err.h"
 
-/* ----- VOLUME ----- */
-
-struct Volume {
-    struct Device* dev;
-
-    SECTNUM firstBlock;     /* first block of data area (from beginning of device) */
-    SECTNUM lastBlock;      /* last block of data area  (from beginning of device) */
-    SECTNUM rootBlock;      /* root block (from firstBlock) */
-
-    char dosType;           /* FFS/OFS, DIRCACHE, INTERNATIONAL */
-    BOOL bootCode;
-    BOOL readOnly;
-    int datablockSize;      /* 488 or 512 */
-    int blockSize;			/* 512 */
-
-    char *volName;
-
-    BOOL mounted;
-
-    int32_t bitmapSize;             /* in blocks */
-    SECTNUM *bitmapBlocks;       /* bitmap blocks pointers */
-    struct bBitmapBlock **bitmapTable;
-    BOOL *bitmapBlocksChg;
-
-    SECTNUM curDirPtr;
-};
 
 
 struct Partition {
